Restructure the tests to make them easier to understand.
Originally we had a bunch of JSON files that we'd use as the input and expected output of compileAST. This was because the output of generate() was troublesome to predict and to write as it tended to contain a lot more newlines than normal for short expressions. Now we normalize the "expected" JS source by parsing it and re-generating it so that it will match the "actual" if their ASTs are the same. This prevents annoying whitespace differences.